+++ src/lib.rs
@@ -5,6 +5,7 @@
 #![doc = include_str!("../README.md")]
 
 mod ipnet;
+mod multicast;
 #[cfg(feature = "schemars")]
 mod schema_util;
 
@@ -12,3 +13,4 @@ pub use ipnet::{
     IpNet, IpNetParseError, IpNetPrefixError, Ipv4Net, Ipv6Net, IPV4_NET_WIDTH_MAX,
     IPV6_NET_WIDTH_MAX,
 };
+pub use multicast::MulticastMac;